feat: update EXECUTION

+++ springboot-app/src/main/java/com/lumos/lumosspring/stock/entities/Material.java
@@ -147,7 +147,7 @@ public class Material {
         this.stockQuantity -= quantityCompleted;
     }
 
-    public void removeStockAvailable(float quantityAvailable) {
+    public void removeStockAvailable(double quantityAvailable) {
         this.stockAvailable -= quantityAvailable;
     }
